chore(auth): rename logout_test.go to handler_test.go; merge logout_redirect_test.go into handler_test.go
