+++ internal/auth/handler_test.go
@@ -10,6 +10,18 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
+func TestLogoutHandler_StatusAndLocation(t *testing.T) {
+	req := httptest.NewRequest(http.MethodGet, "/logout", nil)
+	w := httptest.NewRecorder()
+
+	LogoutHandler(w, req)
+
+	resp := w.Result()
+
+	assert.Equal(t, http.StatusFound, resp.StatusCode)
+	assert.Equal(t, "https://stab.com/logout", resp.Header.Get("Location"))
+}
+
 func TestMetadataHandler_ReturnsExpectedStatusAndJSONBody(t *testing.T) {
 	req := httptest.NewRequest(http.MethodGet, "/logout/redirect", nil)
 	w := httptest.NewRecorder()
