--- internal/auth/logout_test.go
@@ -1,21 +1,0 @@
-package auth
-
-import (
-	"net/http"
-	"net/http/httptest"
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-)
-
-func TestLogoutHandler_StatusAndLocation(t *testing.T) {
-	req := httptest.NewRequest(http.MethodGet, "/logout", nil)
-	w := httptest.NewRecorder()
-
-	LogoutHandler(w, req)
-
-	resp := w.Result()
-
-	assert.Equal(t, http.StatusFound, resp.StatusCode)
-	assert.Equal(t, "https://stab.com/logout", resp.Header.Get("Location"))
-}
